Deprecate /appointments end point (#209)
This end point was created before we had group bookings, so it doesn't return group bookings. This is unexpected for new users and since we already have other end points for individual appointment, group appointment, unavailable blocks and bookings, we are removing this from the documentation although we are still going to support it.